backend/services/recommend/requirements.lambda.txt
# Recommend Service - Lambda Requirements
fastapi==0.109.0
mangum==0.17.0
pydantic-settings==2.1.0
httpx==0.27.0
boto3==1.34.0
six
pymysql
slowapi
google-generativeai
google-api-python-client
elasticsearch==8.12.0
